// src/app/shared/generated/model/getParametersRequest.ts
/**
 * onecx-shell-bff
 * This API provides endpoints to load essential data like workspace configurations for the shell
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface GetParametersRequest { 
    products: { [key: string]: Array<string>; };
}